更新了依赖包，可以使用较新版的各个依赖包，如torch，numpy，matplotlib，scikit-image，tqdm，tensorboard，opencv-python等，修复了一些文件夹未创建的问题，增加了.gitignore文件。

+++ src/Base/Display.py
@@ -57,7 +57,7 @@ class Display:
 
         neither_color = np.ones(
             (area_y_max, area_x_max, 3),
-            dtype=np.float)  # 设置既不是禁飞区、也不是起飞区、也不是green参数指定区域的部分为白色
+            dtype=np.float32)  # 设置既不是禁飞区、也不是起飞区、也不是green参数指定区域的部分为白色
         grid_image = green_color * green + nfz_color * nfz + slz_color * slz + neither_color * neither  # 合成网格图像
 
         # 根据value_map来转换网格图片的明亮度
